Catch case when a timestamp has not yet been set
In this case the function will print 'never' instead of the
relative time since the latest timestamp
